services/aggregate/requirements.txt
pymysql
python-dotenv
sqlalchemy
pandas
